Simple automount script for USB sticks
https://github.com/Sharrisii/TAZ/issues/4
Try simple approach of script + udev rule to mount USB sticks
Assumes the first partition is the one of interest.

The current implementation won't work in the case of a stick that was
directly formatted as filesystem without any partition table, which
is something I have occasionally run across.
Do we care about that case?