Add simple xorg.conf with fbdev as driver
This xorg.conf picks fbdev driver; kms is now the default for xorg 1.19
if no config file is found.

It might be good to switch to kms, but not sure how well it will do with
the older hw we are trying to support. It would need some testing first.